pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "babelbit"
version = "0.1.0"
description = "Babelbit CLI and tools"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Babelbit" }]

# Core runtime deps
dependencies = [
  "substrate-interface",
  "pydantic>=2",
  "httpx",
  "python-dotenv>=0.21.0",
  "aiohttp>=3.9",
  "Pillow>=10.0",
  "opencv-python>=4.8",
  "click>=8.0.0",
  "huggingface_hub>=0.24",
  "bittensor",
  "jinja2>=3.1.6",
  "chutes>=0.3.33",
  "torch",
  "aiobotocore==2.13.1",
  "numpy>=1.24",
  "pynacl>=1.5",
  "fastapi",
  "uvicorn",
  "petname",
  "transformers>=4.56.0",
  "requests>=2.32.5",
  "asyncpg>=0.29.0",
  "boto3>=1.34.131",
  "openai>=2.1.0",
  "dotenv>=0.9.9",
]
 
[project.scripts]
bb = "babelbit:cli"

[project.optional-dependencies]
# Heavy ML stack used by chute templates; optional to keep base lean.
ml = [
  "torch",
  "torchvision",
  "torchaudio",
  "transformers",
]

[tool.hatch.build.targets.wheel]
packages = ["babelbit"]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "httpx>=0.24.0",
    "testcontainers[postgresql]>=4.8.0",
]

